django/pyproject.toml
[project]
name = "student-attendance-management"
version = "0.1.0"
description = "Student Attendance Management System with Face Recognition"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14",
    "django-cors-headers>=4.3",
    "djangorestframework-simplejwt>=5.3",
    "python-dotenv>=1.0.1",
    "astroid==3.2.3",
    "click==8.1.7",
    "cmake==3.30.0",
    "colorama==0.4.6",
    "dill==0.3.8",
    "dill==0.3.8",
    # "dlib",  # Manually managed due to build issues
    # "face-recognition", # Manually managed due to build issues
    "face_recognition_models==0.3.0",
    "isort==5.13.2",
    "mccabe==0.7.0",
    "numpy==1.26.4",
    "opencv-python==4.10.0.84",
    "pillow==10.4.0",
    "platformdirs==4.2.2",
    "pylint==3.2.5",
    "setuptools==70.3.0",
    "tomlkit==0.13.0",
    "drf-spectacular>=0.27.0",
]

[tool.uv]
package = false

[tool.setuptools]
packages = ["accounts", "attendance", "config"]

[dependency-groups]
dev = []
